# Optimize HTTP method lookup in routing jumptable