--- a/src/Http/Routing/perf/Microbenchmarks/Matching/HttpMethodPolicyJumpTableBenchmark.cs
+++ b/src/Http/Routing/perf/Microbenchmarks/Matching/HttpMethodPolicyJumpTableBenchmark.cs
@@ -11,32 +11,32 @@ public class HttpMethodPolicyJumpTableBenchmark
     private PolicyJumpTable _dictionaryJumptable;
     private PolicyJumpTable _singleEntryJumptable;
     private DefaultHttpContext _httpContext;
+    private Dictionary<string, int> _destinations = new();
+    private Dictionary<string, int> _corsDestinations = new();
+
+    private static string[] AllHttpMethods = [HttpMethods.Get, HttpMethods.Connect, HttpMethods.Delete, HttpMethods.Head, HttpMethods.Options, HttpMethods.Patch, HttpMethods.Put, HttpMethods.Post, HttpMethods.Trace];
+
+    [Params(2, 4, 9)]
+    public int DestinationCount { get; set; }
 
     [GlobalSetup]
     public void Setup()
     {
-        _dictionaryJumptable = new HttpMethodDictionaryPolicyJumpTable(
-            0,
-            new Dictionary<string, int>
-            {
-                [HttpMethods.Get] = 1
-            },
-            -1,
-            new Dictionary<string, int>
-            {
-                [HttpMethods.Get] = 2
-            });
-        _singleEntryJumptable = new HttpMethodSingleEntryPolicyJumpTable(
-            0,
-            HttpMethods.Get,
-            -1,
-            supportsCorsPreflight: true,
-            -1,
-            2);
+        for (int i = 0; i < DestinationCount; i++)
+        {
+            _destinations.Add(AllHttpMethods[i], i);
+            _corsDestinations.Add(AllHttpMethods[i], DestinationCount + i);
+        }
 
+        _dictionaryJumptable = new HttpMethodDictionaryPolicyJumpTable(0, _destinations, -1, _corsDestinations);
+        _singleEntryJumptable = new HttpMethodSingleEntryPolicyJumpTable(0, HttpMethods.Get, -1, supportsCorsPreflight: true, -1, 2);
         _httpContext = new DefaultHttpContext();
         _httpContext.Request.Method = HttpMethods.Get;
     }
+
+    [Benchmark]
+    public PolicyJumpTable DictionaryPolicyJumpTableCtor() =>
+        new HttpMethodDictionaryPolicyJumpTable(0, _destinations, -1, _corsDestinations);
 
     [Benchmark]
     public int DictionaryPolicyJumpTable()