--- a/src/Http/Routing/perf/Microbenchmarks/Matching/HttpMethodPolicyJumpTableBenchmark.cs
+++ b/src/Http/Routing/perf/Microbenchmarks/Matching/HttpMethodPolicyJumpTableBenchmark.cs
@@ -12,7 +12,6 @@ public class HttpMethodPolicyJumpTableBenchmark
     private PolicyJumpTable _singleEntryJumptable;
     private DefaultHttpContext _httpContext;
     private Dictionary<string, int> _destinations = new();
-    private Dictionary<string, int> _corsDestinations = new();
 
     [Params("GET", "POST", "Merge")]
     public string TestHttpMethod { get; set; }
@@ -20,24 +19,36 @@ public class HttpMethodPolicyJumpTableBenchmark
     [GlobalSetup]
     public void Setup()
     {
-        var knownJumpTable = new KnownHttpMethodsJumpTable()
-        {
-            ConnectDestination = 1,
-            DeleteDestination = 2,
-            HeadDestination = 3,
-            GetDestination = 4,
-            OptionsDestination = 5,
-            PatchDestination = 6,
-            PutDestination = 7,
-            PostDestination = 8,
-            TraceDestination = 9
-        };
         _destinations.Add("MERGE", 10);
+        var lookup = CreateLookup(_destinations);
 
-        _dictionaryJumptable = new HttpMethodDictionaryPolicyJumpTable(0, knownJumpTable, _destinations, supportsCorsPreflight: false, -1, knownJumpTable, _corsDestinations);
+        _dictionaryJumptable = new HttpMethodDictionaryPolicyJumpTable(lookup, corsPreflightDestinations: null);
         _singleEntryJumptable = new HttpMethodSingleEntryPolicyJumpTable(0, HttpMethods.Get, -1, supportsCorsPreflight: false, -1, 2);
         _httpContext = new DefaultHttpContext();
         _httpContext.Request.Method = TestHttpMethod;
+    }
+
+    private static HttpMethodDestinationsLookup CreateLookup(Dictionary<string, int> extra)
+    {
+        var destinations = new List<KeyValuePair<string, int>>
+        {
+            KeyValuePair.Create(HttpMethods.Connect, 1),
+            KeyValuePair.Create(HttpMethods.Delete, 2),
+            KeyValuePair.Create(HttpMethods.Head, 3),
+            KeyValuePair.Create(HttpMethods.Get, 4),
+            KeyValuePair.Create(HttpMethods.Options, 5),
+            KeyValuePair.Create(HttpMethods.Patch, 6),
+            KeyValuePair.Create(HttpMethods.Put, 7),
+            KeyValuePair.Create(HttpMethods.Post, 8),
+            KeyValuePair.Create(HttpMethods.Trace, 9)
+        };
+
+        foreach (var item in extra)
+        {
+            destinations.Add(item);
+        }
+
+        return new HttpMethodDestinationsLookup(destinations, exitDestination: 0);
     }
 
     [Benchmark]