--- a/src/Http/Routing/src/Matching/HttpMethodMatcherPolicy.cs
+++ b/src/Http/Routing/src/Matching/HttpMethodMatcherPolicy.cs
@@ -533,10 +533,6 @@ public sealed class HttpMethodMatcherPolicy : MatcherPolicy, IEndpointComparerPo
         private const int PutIndex = 7;
         private const int TraceIndex = 8;
 
-        public KnownHttpMethodsJumpTableBuilder()
-        {
-        }
-
         private BuildTable _table;
 
         public bool TrySet(string method, int destination)