# (Not?) Using FrozenDictionary in HttpMethodDictionaryPolicyJumpTable

Commit 1-2:
- Converting Dictionary to Frozen in the ctor of HttpMethodDictionaryPolicyJumpTable.
- Adding performance tests for the ctor.

After discussions Commit 3:
- Based on performance measurement commit 3 introduces knowledge on known HTTP methods and removes frozen dictionary. It is using a switch expression (like .NET runtime) to match the known verbs without allocation.

## Description

My performance measurements indicate a quite significant increase in the startup time as well as allocations (vs the decrease in DictionaryPolicyJumpTable which is the goal of the PR).

Before:
```
// * Summary *

BenchmarkDotNet=v0.13.0, OS=Windows 10.0.22621
12th Gen Intel Core i7-1255U, 1 CPU, 12 logical and 10 physical cores
.NET SDK=9.0.100-alpha.1.23524.8
  [Host]     : .NET 9.0.0 (9.0.23.52802), X64 RyuJIT
  Job-BZJOQC : .NET 9.0.0 (9.0.23.52313), X64 RyuJIT

Server=True  Toolchain=.NET Core 9.0  RunStrategy=Throughput

|                        Method | DestinationCount |     Mean |     Error |    StdDev |          Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------ |----------------- |---------:|----------:|----------:|--------------:|-------:|------:|------:|----------:|
| DictionaryPolicyJumpTableCtor |                2 | 8.300 ns | 0.0806 ns | 0.0715 ns | 120,475,250.4 | 0.0005 |     - |     - |      48 B |
|     DictionaryPolicyJumpTable |                2 | 9.422 ns | 0.0843 ns | 0.0747 ns | 106,140,068.2 |      - |     - |     - |         - |
|    SingleEntryPolicyJumpTable |                2 | 4.637 ns | 0.0964 ns | 0.1382 ns | 215,642,963.9 |      - |     - |     - |         - |
| DictionaryPolicyJumpTableCtor |                4 | 8.117 ns | 0.0531 ns | 0.0497 ns | 123,195,634.7 | 0.0005 |     - |     - |      48 B |
|     DictionaryPolicyJumpTable |                4 | 9.397 ns | 0.0720 ns | 0.0673 ns | 106,415,958.1 |      - |     - |     - |         - |
|    SingleEntryPolicyJumpTable |                4 | 4.526 ns | 0.0785 ns | 0.0696 ns | 220,966,690.7 |      - |     - |     - |         - |
| DictionaryPolicyJumpTableCtor |                9 | 8.232 ns | 0.0776 ns | 0.0726 ns | 121,480,877.5 | 0.0005 |     - |     - |      48 B |
|     DictionaryPolicyJumpTable |                9 | 9.494 ns | 0.0836 ns | 0.0653 ns | 105,328,310.6 |      - |     - |     - |         - |
|    SingleEntryPolicyJumpTable |                9 | 4.435 ns | 0.0368 ns | 0.0326 ns | 225,503,508.5 |      - |     - |     - |         - |
```

After (using FrozenDictionary):

```
|                        Method | DestinationCount |       Mean |     Error |    StdDev |          Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------ |----------------- |-----------:|----------:|----------:|--------------:|-------:|------:|------:|----------:|
| DictionaryPolicyJumpTableCtor |                2 | 390.895 ns | 5.2786 ns | 4.6794 ns |   2,558,230.8 | 0.0129 |     - |     - |   1,200 B |
|     DictionaryPolicyJumpTable |                2 |   6.799 ns | 0.0935 ns | 0.0829 ns | 147,087,596.4 |      - |     - |     - |         - |
|    SingleEntryPolicyJumpTable |                2 |   4.466 ns | 0.0404 ns | 0.0338 ns | 223,917,894.9 |      - |     - |     - |         - |
| DictionaryPolicyJumpTableCtor |                4 | 496.404 ns | 6.9354 ns | 6.4874 ns |   2,014,489.3 | 0.0153 |     - |     - |   1,472 B |
|     DictionaryPolicyJumpTable |                4 |   6.786 ns | 0.0474 ns | 0.0396 ns | 147,355,032.1 |      - |     - |     - |         - |
|    SingleEntryPolicyJumpTable |                4 |   4.467 ns | 0.0423 ns | 0.0375 ns | 223,866,735.8 |      - |     - |     - |         - |
| DictionaryPolicyJumpTableCtor |                9 | 771.634 ns | 3.9767 ns | 3.3207 ns |   1,295,951.9 | 0.0191 |     - |     - |   1,824 B |
|     DictionaryPolicyJumpTable |                9 |   6.789 ns | 0.0456 ns | 0.0426 ns | 147,288,554.2 |      - |     - |     - |         - |
|    SingleEntryPolicyJumpTable |                9 |   4.535 ns | 0.0375 ns | 0.0332 ns | 220,523,926.8 |      - |     - |     - |         - |
```

It is unclear for me if this change brings enough value to be merged. 

After (using switch expression):

```
|                        Method | DestinationCount |      Mean |     Error |    StdDev |          Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------ |----------------- |----------:|----------:|----------:|--------------:|-------:|------:|------:|----------:|
| DictionaryPolicyJumpTableCtor |                2 | 27.427 ns | 0.2116 ns | 0.1875 ns |  36,460,073.0 | 0.0013 |     - |     - |     120 B |
|     DictionaryPolicyJumpTable |                2 |  5.771 ns | 0.0580 ns | 0.0514 ns | 173,274,603.3 |      - |     - |     - |         - |
|    SingleEntryPolicyJumpTable |                2 |  4.408 ns | 0.0460 ns | 0.0431 ns | 226,851,759.1 |      - |     - |     - |         - |
| DictionaryPolicyJumpTableCtor |                4 | 41.707 ns | 0.8466 ns | 0.9410 ns |  23,976,899.9 | 0.0013 |     - |     - |     120 B |
|     DictionaryPolicyJumpTable |                4 |  5.663 ns | 0.0332 ns | 0.0310 ns | 176,582,577.0 |      - |     - |     - |         - |
|    SingleEntryPolicyJumpTable |                4 |  4.361 ns | 0.0197 ns | 0.0175 ns | 229,311,321.4 |      - |     - |     - |         - |
| DictionaryPolicyJumpTableCtor |                9 | 91.236 ns | 0.4729 ns | 0.4192 ns |  10,960,590.1 | 0.0012 |     - |     - |     120 B |
|     DictionaryPolicyJumpTable |                9 |  5.637 ns | 0.0176 ns | 0.0138 ns | 177,390,232.5 |      - |     - |     - |         - |
|    SingleEntryPolicyJumpTable |                9 |  4.366 ns | 0.0315 ns | 0.0295 ns | 229,050,294.5 |      - |     - |     - |         - |
```

cc. @BrennanConroy 

Fixes #47569