@ladeak 2023-11-01
@BrennanConroy I am not sure if the change worth merging based on the significant tradeoff at startup time. 

Does `HttpMethodDictionaryPolicyJumpTable` only allow HTTP method as the key of the lookup table? The type of the key is `string`, and currently it does not restrict to the known http verbs. I am asking to understand if it would be possible to use a different data type (ie. an array with a switch expression), given the HTTP verbs are well-known. Or does this type need to handle cases that are not known http verbs?