@BrennanConroy 2023-11-02
> EDIT: thinking about it further, HttpMethodDictionaryPolicyJumpTable could handle the known HTTP verbs with a switch while the unknown ones could be handled similarly as today with a Dictionary. With such for known HTTP methods we could achieve the below results, but note that I have not yet pushed this commit.

So you almost doubled the perf by using a switch instead? Whereas frozen dictionary "only" got ~50% faster?