@ladeak 2023-11-02
@BrennanConroy it was simply using 9 `if(string.Equals(...,..., OrdinalIgnoreCase))` for the known values and dictionary fallback for the unknown values. Using switch expression is trickier because of case insesitiveness, I would need to stackalloc to avoid extra allocation (stackalloc with upper limit of the longest known value). 
I will take a look at the suggested PR, and measure both and come back to this thread.