@ladeak 2023-11-02
Using switch-expressions as committed.

```
|                        Method | DestinationCount |      Mean |     Error |    StdDev |          Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------ |----------------- |----------:|----------:|----------:|--------------:|-------:|------:|------:|----------:|
| DictionaryPolicyJumpTableCtor |                2 | 27.427 ns | 0.2116 ns | 0.1875 ns |  36,460,073.0 | 0.0013 |     - |     - |     120 B |
|     DictionaryPolicyJumpTable |                2 |  5.771 ns | 0.0580 ns | 0.0514 ns | 173,274,603.3 |      - |     - |     - |         - |
|    SingleEntryPolicyJumpTable |                2 |  4.408 ns | 0.0460 ns | 0.0431 ns | 226,851,759.1 |      - |     - |     - |         - |
| DictionaryPolicyJumpTableCtor |                4 | 41.707 ns | 0.8466 ns | 0.9410 ns |  23,976,899.9 | 0.0013 |     - |     - |     120 B |
|     DictionaryPolicyJumpTable |                4 |  5.663 ns | 0.0332 ns | 0.0310 ns | 176,582,577.0 |      - |     - |     - |         - |
|    SingleEntryPolicyJumpTable |                4 |  4.361 ns | 0.0197 ns | 0.0175 ns | 229,311,321.4 |      - |     - |     - |         - |
| DictionaryPolicyJumpTableCtor |                9 | 91.236 ns | 0.4729 ns | 0.4192 ns |  10,960,590.1 | 0.0012 |     - |     - |     120 B |
|     DictionaryPolicyJumpTable |                9 |  5.637 ns | 0.0176 ns | 0.0138 ns | 177,390,232.5 |      - |     - |     - |         - |
|    SingleEntryPolicyJumpTable |                9 |  4.366 ns | 0.0315 ns | 0.0295 ns | 229,050,294.5 |      - |     - |     - |         - |
```

I would like to re-use somehow the code in Kestrel (https://github.com/dotnet/aspnetcore/pull/44096), but not sure what common place I may move it to, and having reference to Kestrel does not seem right.