@ladeak 2023-11-11
I have updated the PR to take a direction as suggested in previous comments. The "known jump table" is now created in the BuildPolicy method.

 I update the benchmarks here. However, please note that I have switched off the support preflight flag, hence a slight improvement in all benchmarks below compared to the ones previously commented on in this post.

### HttpMethodPolicyJumpTableBenchmark

#### BEFORE

For transparency (and because the ctor `HttpMethodDictionaryPolicyJumpTable` has changed) this is the `Setup` method of `HttpMethodPolicyJumpTableBenchmark` I used to measure the *before* performance. The *after* one is the one available in the PR.

```csharp
[GlobalSetup]
public void Setup()
{
    _destinations.Add(HttpMethods.Connect, 1);
    _destinations.Add(HttpMethods.Delete, 2);
    _destinations.Add(HttpMethods.Head, 3);
    _destinations.Add(HttpMethods.Get, 4);
    _destinations.Add(HttpMethods.Options, 5);
    _destinations.Add(HttpMethods.Patch, 6);
    _destinations.Add(HttpMethods.Post, 7);
    _destinations.Add(HttpMethods.Put, 8);
    _destinations.Add(HttpMethods.Trace, 9);
    _destinations.Add("MERGE", 10);

    _dictionaryJumptable = new HttpMethodDictionaryPolicyJumpTable(0, _destinations, -1, _corsDestinations);
    _singleEntryJumptable = new HttpMethodSingleEntryPolicyJumpTable(0, HttpMethods.Get, -1, supportsCorsPreflight: false, -1, 2);
    _httpContext = new DefaultHttpContext();
    _httpContext.Request.Method = TestHttpMethod;
}
```

```
|                     Method | TestHttpMethod |     Mean |     Error |    StdDev |   Median |          Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------- |--------------- |---------:|----------:|----------:|---------:|--------------:|------:|------:|------:|----------:|
|  DictionaryPolicyJumpTable |            GET | 8.338 ns | 0.0704 ns | 0.0658 ns | 8.341 ns | 119,936,347.4 |     - |     - |     - |         - |
| SingleEntryPolicyJumpTable |            GET | 3.108 ns | 0.0225 ns | 0.0211 ns | 3.101 ns | 321,710,436.1 |     - |     - |     - |         - |
|  DictionaryPolicyJumpTable |          Merge | 8.341 ns | 0.0575 ns | 0.0538 ns | 8.344 ns | 119,893,341.5 |     - |     - |     - |         - |
| SingleEntryPolicyJumpTable |          Merge | 3.615 ns | 0.0768 ns | 0.1479 ns | 3.680 ns | 276,609,311.1 |     - |     - |     - |         - |
|  DictionaryPolicyJumpTable |           POST | 8.290 ns | 0.0912 ns | 0.0809 ns | 8.283 ns | 120,633,371.2 |     - |     - |     - |         - |
| SingleEntryPolicyJumpTable |           POST | 3.381 ns | 0.0556 ns | 0.0493 ns | 3.376 ns | 295,737,987.0 |     - |     - |     - |         - |
```

#### AFTER

```
|                     Method | TestHttpMethod |     Mean |     Error |    StdDev |          Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------- |--------------- |---------:|----------:|----------:|--------------:|------:|------:|------:|----------:|
|  DictionaryPolicyJumpTable |            GET | 3.807 ns | 0.0791 ns | 0.0739 ns | 262,683,263.4 |     - |     - |     - |         - |
| SingleEntryPolicyJumpTable |            GET | 3.089 ns | 0.0116 ns | 0.0090 ns | 323,756,224.3 |     - |     - |     - |         - |
|  DictionaryPolicyJumpTable |          Merge | 8.800 ns | 0.0454 ns | 0.0402 ns | 113,641,755.1 |     - |     - |     - |         - |
| SingleEntryPolicyJumpTable |          Merge | 3.359 ns | 0.0210 ns | 0.0196 ns | 297,748,718.0 |     - |     - |     - |         - |
|  DictionaryPolicyJumpTable |           POST | 3.743 ns | 0.0209 ns | 0.0186 ns | 267,176,079.6 |     - |     - |     - |         - |
| SingleEntryPolicyJumpTable |           POST | 3.381 ns | 0.0300 ns | 0.0280 ns | 295,781,095.6 |     - |     - |     - |         - |
```



### HttpMethodMatcherPolicyBenchmark

#### BEFORE

```
|         Method | DestinationCount |      Mean |    Error |   StdDev |         Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------- |----------------- |----------:|---------:|---------:|-------------:|-------:|------:|------:|----------:|
| BuildJumpTable |                3 |  96.43 ns | 1.016 ns | 0.848 ns | 10,370,136.0 | 0.0042 |     - |     - |     392 B |
| BuildJumpTable |                5 | 104.21 ns | 1.723 ns | 1.439 ns |  9,596,231.4 | 0.0042 |     - |     - |     392 B |
| BuildJumpTable |               11 | 149.29 ns | 2.491 ns | 2.330 ns |  6,698,388.4 | 0.0041 |     - |     - |     392 B |
```

#### AFTER

```
|         Method | DestinationCount |      Mean |    Error |   StdDev |         Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------- |----------------- |----------:|---------:|---------:|-------------:|-------:|------:|------:|----------:|
| BuildJumpTable |                3 |  83.29 ns | 0.551 ns | 0.430 ns | 12,005,749.0 | 0.0029 |     - |     - |     264 B |
| BuildJumpTable |                5 | 142.14 ns | 1.974 ns | 1.750 ns |  7,035,194.8 | 0.0055 |     - |     - |     512 B |
| BuildJumpTable |               11 | 278.49 ns | 2.404 ns | 2.131 ns |  3,590,761.2 | 0.0110 |     - |     - |   1,040 B |
```

@JamesNK is this the direction you would like to go?